+++ CHANGELOG.md
@@ -5,6 +5,9 @@ using the [Keep a CHANGELOG](https://keepachangelog.com/) principles.
 
 ## [UNRELEASED]
 
+### Changed
+
+- Removed dependency **symfony/yaml** for less conflicts with composer
 - DirectoryExists Tests does now throw an error if no directory was specified.
 - FileExists Tests does now throw an error if no file was specified.
 - FileContentTest Tests does now throw an error if no file or expected/unexpected conditions where specified.