+++ CHANGELOG.md
@@ -7,7 +7,7 @@ using the [Keep a CHANGELOG](https://keepachangelog.com/) principles.
 
 ### Changed
 
-- Removed dependency **symfony/yaml** for less conflicts with composer
+- Change dependencies of symfony/yaml and twig/twig to * for better composer compatibility in other projects
 - DirectoryExists Tests does now throw an error if no directory was specified.
 - FileExists Tests does now throw an error if no file was specified.
 - FileContentTest Tests does now throw an error if no file or expected/unexpected conditions where specified.