+++ CHANGELOG.md
@@ -15,7 +15,8 @@ using the [Keep a CHANGELOG](https://keepachangelog.com/) principles.
 - DirectoryExists Tests does now throw an error if no directory was specified.
 - FileExists Tests does now throw an error if no file was specified.
 - FileContentTest Tests does now throw an error if no file or expected/unexpected conditions where specified.
-
+- Allow test files with a "*.yaml" extension and not only "*.yml"
+- 
 ## [1.7.0]
 
 ### Added